Refactor buffer into separate subpackage
